requirements.txt
google-genai==0.3.4
supabase==2.7.4
python-dotenv==1.0.1
